Refine strategy implementation split
