+++ master-firmware/package.yml
@@ -65,7 +65,8 @@ target.arm:
     - src/manipulator/gripper.cpp
     - src/manipulator/manipulator_thread.cpp
     - src/uid.c
-    - src/strategy_impl.cpp
+    - src/strategy_impl/actions.cpp
+    - src/strategy_impl/base.cpp
     - src/strategy_simulated.cpp
 
 source:
